pyproject.toml
[project]
name = "marshmallow"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "fire>=0.7.1",
    "google-cloud-storage>=3.10.1",
    "lancedb>=0.30.2",
    "openai>=2.30.0",
    "pypdf>=6.9.2",
    "redis>=7.4.0",
    "uvicorn>=0.34.0",
]

[project.scripts]
lancedb-service = "lancedb_service.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["lancedb_service", "other-service"]
